requirements.txt
langchain==0.3.9
langchain_community==0.3.9
langchain_core==0.3.21
langchain_openai==0.0.5
python-dotenv==1.0.1
streamlit==1.30.0
chromadb==0.3.29
beautifulsoup4==4.12.2
selenium
